# crates/chat-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "scion-chat",
    "description": "A chat server over HTTP/3-over-SCION",
    "contact": {
      "name": "Anapaya & contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/login": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Exchange a username and password for a bearer token.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The token to send as `Authorization: Bearer`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "The username and password do not match",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/register": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Register an account.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The account was created"
          },
          "409": {
            "description": "The username is taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The username is not acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "The server accepts no more accounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rooms/{id}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "Read a page of a room's messages, oldest first.",
        "operationId": "list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The room to read",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many messages to return, clamped to 1..=200. Defaults to 50.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "description": "Return only messages newer than this `seq`.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "before_seq",
            "in": "query",
            "description": "Return only messages older than this `seq`.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of messages, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessagesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Both cursors were given",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No usable bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No room with that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Append a message to a room, attributed to the token holder.",
        "operationId": "post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The room to post to",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Where the message landed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "No usable bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No room with that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "The body is larger than the server accepts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The body is empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "The contents of an [`ErrorResponse`].",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What failed, as a value a client can branch on. The HTTP status alone is too coarse."
          },
          "message": {
            "type": "string",
            "description": "A human-readable explanation, for logs and for showing to a user. Free-form: it may change\nbetween server versions, so never branch on it."
          }
        }
      },
...
      "ErrorCode": {
        "oneOf": [
          {
            "type": "string",
            "description": "The request body was missing, malformed, or not the shape the endpoint expects.",
            "enum": [
              "invalid_body"
            ]
          },
          {
            "type": "string",
            "description": "The server failed for a reason it does not describe further.",
            "enum": [
              "internal"
            ]
          },
          {
            "type": "string",
            "description": "The request carried no usable bearer token.",
            "enum": [
              "unauthorized"
            ]
          },
          {
            "type": "string",
            "description": "The username exists but the password does not match it.",
            "enum": [
              "invalid_credentials"
            ]
          },
          {
            "type": "string",
            "description": "The requested username is already registered.",
            "enum": [
              "username_taken"
            ]
          },
          {
            "type": "string",
            "description": "The requested username is outside what registration accepts.",
            "enum": [
              "invalid_username"
            ]
          },
          {
            "type": "string",
            "description": "The requested room name is outside what room creation accepts.",
            "enum": [
              "invalid_name"
            ]
          },
          {
            "type": "string",
            "description": "No room has the requested id.",
            "enum": [
              "room_not_found"
            ]
          },
          {
            "type": "string",
            "description": "The message body exceeds the server's limit.",
            "enum": [
              "message_too_large"
            ]
          },
          {
            "type": "string",
            "description": "A configured cap — accounts or rooms — is already reached.",
            "enum": [
              "cap_exceeded"
            ]
          },
          {
            "type": "object",
            "description": "A code this build does not know, kept verbatim.\n\nDecoding an error must not itself fail: a client that meets a code added after it was\nbuilt still surfaces the server's message rather than a parse error.",
            "required": [
              "other"
            ],
            "properties": {
              "other": {
                "$ref": "#/components/schemas/UnknownCode",
                "description": "A code this build does not know, kept verbatim.\n\nDecoding an error must not itself fail: a client that meets a code added after it was\nbuilt still surfaces the server's message rather than a parse error."
              }
            }
          }
        ],
        "description": "Every failure this API distinguishes, on the wire as its name in `snake_case`.\n\nAn enum rather than a string so the server cannot emit a code that does not exist and a client\ncan `match` with the compiler checking it has covered them all."
      },
      "ErrorResponse": {
        "type": "object",
        "description": "The body of every failing response, whatever failed.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError",
            "description": "What went wrong."
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "The credentials an existing account is authenticated with.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "The password in the clear, verified against the stored hash."
          },
          "username": {
            "type": "string",
            "description": "The registered name to log in as."
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "description": "What a successful login yields: the bearer token every authenticated request carries.",
        "required": [
          "token",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "$ref": "#/components/schemas/UnixMillis",
            "description": "When the token stops being accepted. There are no refresh tokens — a client logs in again."
          },
          "token": {
            "type": "string",
            "description": "The JWT to send as `Authorization: Bearer <token>`. Opaque to clients: they carry it, they\ndo not parse it."
          }
        }
      },
      "Message": {
        "type": "object",
        "description": "One message in a room.\n\nMessages are append-only: never edited, never deleted.",
        "required": [
          "seq",
          "username",
          "body",
          "posted_at"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The message text."
          },
          "posted_at": {
            "$ref": "#/components/schemas/UnixMillis",
            "description": "When the server accepted the message."
          },
          "seq": {
            "$ref": "#/components/schemas/Seq",
            "description": "The message's position in the sequence, and the cursor used to fetch around it."
          },
          "username": {
            "type": "string",
            "description": "The account that posted it."
          }
        }
      },
      "MessagesResponse": {
        "type": "object",
        "description": "A page of a room's messages.",
        "required": [
          "messages"
        ],
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "The page, **always oldest-first**, whichever cursor asked for it.\n\nA page shorter than the requested `limit` means the client has reached the end it was\nwalking towards: the present when polling forwards, the start of history when loading\nolder messages. A full page means more might be waiting — ask again immediately."
          }
        }
      },
      "PostMessageRequest": {
        "type": "object",
        "description": "The text of a message being posted.",
        "required": [
          "body"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The message text. UTF-8, at most the server's `max_message_bytes` when encoded — see\n[`ServerInfo`](super::ServerInfo)."
          }
        }
      },
      "PostMessageResponse": {
        "type": "object",
        "description": "Where a posted message landed.\n\nA client does *not* advance its poll cursor to this `seq` — a message with a lower `seq` may\nstill be waiting to be polled. The cursor only ever follows what polling actually delivered.",
        "required": [
          "seq",
          "posted_at"
        ],
        "properties": {
          "posted_at": {
            "$ref": "#/components/schemas/UnixMillis",
            "description": "When the server accepted the message. Server time — a client's own clock never appears on\nthe wire."
          },
          "seq": {
            "$ref": "#/components/schemas/Seq",
            "description": "The `seq` the server assigned to the new message."
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "description": "The credentials a new account is created with.\n\nThe username is the account's permanent identity — the name messages are attributed to. There\nis no rename, no password change, and no password reset.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "The password in the clear — the connection's TLS is what protects it. The server keeps\nonly a KDF hash."
          },
          "username": {
            "type": "string",
            "description": "The name to register. UTF-8, 1–32 characters, no control characters; compared\ncase-insensitively against the names already taken."
          }
        }
      },
...
      "Seq": {
        "type": "integer",
        "format": "int64",
        "description": "Position of a message in the server's message sequence.\n\n`seq` increases strictly and never regresses, so it is the one total order that every client\nagrees on. It is assigned server-wide rather than per room, so a single room's messages carry\ngaps: treat `seq` as a cursor, never as a count. Clients remember the highest `seq` they have\nseen in a room and poll for what came after it.\n\nNumbering starts at 1, so [`Seq::START`] is the position before every message — an empty room\nand a client that has read nothing both sit there, and asking for what follows it asks for\neverything.",
        "minimum": 0
      },
...
      "UnixMillis": {
        "type": "integer",
        "format": "int64",
        "description": "A point in time on the wire: milliseconds since the Unix epoch, UTC.",
        "minimum": 0
      },
      "UnknownCode": {
        "type": "string",
        "description": "A code carried by a response but not named by [`ErrorCode`].\n\nIts contents are private and it has no constructor, so decoding is the only thing that can\nproduce one. That is what makes a sender structurally unable to invent a code: reaching for\n[`ErrorCode::Other`] instead of adding a variant is not an option it has."
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}